Complete WebGPU integration refactoring and fix test failures
  - Remove unnecessary WebGPUDevice abstraction layer
  - Update all GPU components to use WebGPUContext directly (VoxelGPUManager, GPUOctreeBuilder, GPUVoxelizer)
  - Add missing compute pipeline methods to WebGPUContext
  - Fix RuntimeMemoryManager test failures (memory pressure eviction and node persistence)
  - Fix WGSL shader validation error in ray_traversal.wgsl
  - Fix WebGPU hardware detection to properly initialize and detect available GPUs
  - Add comprehensive test status documentation
  - Archive completed fix documentation

+++ render/src/main/resources/shaders/rendering/ray_traversal.wgsl
@@ -95,7 +95,7 @@ fn traverse_hierarchical(ray: Ray, hit: ptr<function, Hit>) {
         let node = voxel_octree[node_index];
         
         // Calculate node bounds (simplified for uniform octree)
-        let level_size = f32(1 << (config.octree_max_depth - i32(node.level)));
+        let level_size = f32(1u << u32(config.octree_max_depth - i32(node.level)));
         let node_pos = decode_morton_position(node_index, i32(node.level));
         let node_min = node_pos * level_size;
         let node_max = node_min + vec3<f32>(level_size);
